refactors db directories and exports models
