+++ server/db/index.js
@@ -1,15 +1,5 @@
-"use strict";
+const db = require('./db');
 
-const chalk = require("chalk");
-const Sequelize = require("sequelize");
-
-console.log(chalk.yellow("Opening database connection"));
-
-const db = new Sequelize(
-  process.env.DATABASE_URL || `postgres://localhost:5432/chaindora`,
-  {
-    logging: false
-  }
-);
+require('./models');
 
 module.exports = db;
