+++ server/db/models/Song.js
@@ -1,5 +1,5 @@
 const Sequelize = require("sequelize");
-const db = require("../index");
+const db = require("../db");
 
 const Song = db.define("song", {
   title: {
